backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic_core
Pydantic
mangum
yoyo-migrations
alembic
python-dotenv